rename sidebar and add to components/elements/Sidebar
